// frontend/src/generated/models/actionSteeringWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { ActionSteeringWorkspaceResponseActionsItem } from './actionSteeringWorkspaceResponseActionsItem';
import type { ActionSteeringWorkspaceResponseAssurance } from './actionSteeringWorkspaceResponseAssurance';
import type { ActionSteeringWorkspaceResponseBenefitSeriesItem } from './actionSteeringWorkspaceResponseBenefitSeriesItem';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { ActionSteeringWorkspaceResponseDependenciesItem } from './actionSteeringWorkspaceResponseDependenciesItem';
import type { ActionSteeringWorkspaceResponseLineage } from './actionSteeringWorkspaceResponseLineage';
import type { ActionSteeringWorkspaceResponseMetrics } from './actionSteeringWorkspaceResponseMetrics';

export interface ActionSteeringWorkspaceResponse {
  actions?: ActionSteeringWorkspaceResponseActionsItem[];
  as_of: string;
  assurance?: ActionSteeringWorkspaceResponseAssurance;
  benefit_series?: ActionSteeringWorkspaceResponseBenefitSeriesItem[];
  context: WorkspaceContextResponse;
  dependencies?: ActionSteeringWorkspaceResponseDependenciesItem[];
  lineage?: ActionSteeringWorkspaceResponseLineage;
  metrics?: ActionSteeringWorkspaceResponseMetrics;
  projection_version: number;
  source_snapshot_ids?: string[];
}