// frontend/src/generated/models/capitalAllocationWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { CapitalAllocationWorkspaceResponseAllocationItem } from './capitalAllocationWorkspaceResponseAllocationItem';
import type { CapitalAllocationWorkspaceResponseApprovalsItem } from './capitalAllocationWorkspaceResponseApprovalsItem';
import type { CapitalAllocationWorkspaceResponseAssurance } from './capitalAllocationWorkspaceResponseAssurance';
import type { CapitalAllocationWorkspaceResponseCandidatesItem } from './capitalAllocationWorkspaceResponseCandidatesItem';
import type { CapitalAllocationWorkspaceResponseConstraintsItem } from './capitalAllocationWorkspaceResponseConstraintsItem';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { CapitalAllocationWorkspaceResponseFrontierPointsItem } from './capitalAllocationWorkspaceResponseFrontierPointsItem';
import type { CapitalAllocationWorkspaceResponseFundingOptionsItem } from './capitalAllocationWorkspaceResponseFundingOptionsItem';
import type { CapitalAllocationWorkspaceResponseLineage } from './capitalAllocationWorkspaceResponseLineage';
import type { CapitalAllocationWorkspaceResponsePortfolio } from './capitalAllocationWorkspaceResponsePortfolio';
import type { CapitalAllocationWorkspaceResponseSelectedAllocationRunId } from './capitalAllocationWorkspaceResponseSelectedAllocationRunId';

export interface CapitalAllocationWorkspaceResponse {
  allocation?: CapitalAllocationWorkspaceResponseAllocationItem[];
  approvals?: CapitalAllocationWorkspaceResponseApprovalsItem[];
  as_of: string;
  assurance?: CapitalAllocationWorkspaceResponseAssurance;
  candidates?: CapitalAllocationWorkspaceResponseCandidatesItem[];
  constraints?: CapitalAllocationWorkspaceResponseConstraintsItem[];
  context: WorkspaceContextResponse;
  frontier_points?: CapitalAllocationWorkspaceResponseFrontierPointsItem[];
  funding_options?: CapitalAllocationWorkspaceResponseFundingOptionsItem[];
  lineage?: CapitalAllocationWorkspaceResponseLineage;
  portfolio?: CapitalAllocationWorkspaceResponsePortfolio;
  projection_version: number;
  selected_allocation_run_id?: CapitalAllocationWorkspaceResponseSelectedAllocationRunId;
  source_snapshot_ids?: string[];
}